+++ sparql-Loader/src/main/resources/log4j2.xml
